requirements.txt
streamlit
newspaper3k
lxml